+++ src/main/java/com/sanctionco/jmail/disposable/DisposableDomainSource.java
@@ -2,6 +2,8 @@ package com.sanctionco.jmail.disposable;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
 /**
  * <p>A {@code DisposableDomainSource} is used as a source of truth for which domains are considered
@@ -11,9 +13,11 @@ import java.io.InputStream;
  * {@link com.sanctionco.jmail.EmailValidator} in order to consider email addresses that use a
  * disposable domain as invalid.
  *
- * <p>Currently, there are two types of {@code DisposableDomainSource}: {@link FileSource} and
- * {@link IsTempMailAPISource}. Both of these can be instantiated via static methods on this class:
- * {@link #file(String)} and {@link #isTempMailAPI(String)} respectively.
+ * <p>Currently, there are two types of {@code DisposableDomainSource}: {@link InputStreamSource}
+ * and {@link IsTempMailAPISource}. Both of these can be instantiated via static methods on this
+ * class: {@link #inputStream(InputStream)} and {@link #isTempMailAPI(String)} respectively. There
+ * is an additional static method {@link #file(String)} used to instantiate a
+ * {@code DisposableDomainSource} from a file, which uses an {@link InputStreamSource} underneath.
  *
  * <p>Additionally, you can easily create your own {@code DisposableDomainSource} by writing a
  * class that implements {@code DisposableDomainSource}.
@@ -21,7 +25,9 @@ import java.io.InputStream;
 public interface DisposableDomainSource {
 
   /**
-   * Determine if the given domain is a disposable domain.
+   * Determine if the given domain is a disposable domain. Domains are compared in a case
+   * in-sensitive way, so calling this method with both {@code example-disposable-domain.com}
+   * and {@code EXAMPLE-DISPOSABLE-DOMAIN.COM} should return the same result.
    *
    * @param domain the domain to check
    * @return {@code true} if the domain is a disposable domain, or {@code false} if not
@@ -29,18 +35,20 @@ public interface DisposableDomainSource {
   boolean isDisposableDomain(String domain);
 
   /**
-   * <p>Create and return a new {@link FileSource}, which can be used as a
-   * {@code DisposableDomainSource} that uses a file as the source of disposable domains.
+   * <p>Create and return a new {@code DisposableDomainSource} which uses a file as the source
+   * of disposable domains.
    *
    * <p>The file should have each disposable domain on its own line in the file.
    *
    * @param path the path to the file containing disposable domains
-   * @return a new instance of {@link FileSource}
+   * @return a new instance of {@code DisposableDomainSource}
    * @throws IOException if the file at the given path does not exist or there is an issue reading
    *                     the file
    */
   static DisposableDomainSource file(String path) throws IOException {
-    return new FileSource(path);
+    try (InputStream in = Files.newInputStream(Paths.get(path))) {
+      return new InputStreamSource(in);
+    }
   }
 
   /**
@@ -52,7 +60,7 @@ public interface DisposableDomainSource {
    *
    * @param inputStream the input stream containing disposable domains. Will not be closed.
    * @return a new instance of {@link InputStreamSource}
-   * @throws IOException if the input stream is closed or cannot be read
+   * @throws IOException if the input stream is already closed or cannot be read
    */
   static DisposableDomainSource inputStream(InputStream inputStream) throws IOException {
     return new InputStreamSource(inputStream);